committing Tambet's patch:
"Here's a patch to fix the logging output and to implement the progress
bar callback. The previous patch that implemented logging always added a
log callback, so it worked correctly for the first time, but every other
time the output callback was added and a duplicate log line was printed. "