configure sqlite to return utf-8-encoded strs instead of unicode objects
sqlite by default returns all text as unicode objects, and this causes a number of problems when merging repos which
contain utf-8 characters in Provides or Requires (as the current F11/F12 repos do).  For a testcase, try merging 2 F12
repos, and you should see it fail with a UnicodeDecodeError in packages.py:_dump_pco().  This patch instructs sqlite to
return all text as utf-8-encoded strs, which avoids these encoding issues.