setup: make build work on python 3
In python 3, the string module doesn't have 'split' function anymore
and users are encouraged to use '.split()' on the string itself.  Also
'map()' is lazy and returns an iterable map object, but setup expects
only a list, so it has to be wrapped in 'list()'.  All of this works
with python2 as well.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>